ring추가 전 모두 작동
