+++ Cakey/composer/Package.realitycomposerpro/ProjectData/main.json
@@ -1,0 +1,5 @@
+{
+  "pathsToIds" : {
+    "composer\/Sources\/composer\/composer.rkassets\/Scene.usda" : "E804F996-ECB4-423F-895D-FD24E7B88CE4"
+  }
+}
