+++ Cakey/composer/Package.realitycomposerpro/WorkspaceData/dora.rcuserdata
@@ -1,0 +1,238 @@
+{
+  "advancedEditorSelectedIdentifier" : "ProjectBrowser",
+  "openSceneRelativePaths" : [
+    "Scene.usda"
+  ],
+  "recentIdentifiers" : {
+    "E804F996-ECB4-423F-895D-FD24E7B88CE4" : [
+
+    ]
+  },
+  "sceneCameraHistory" : {
+    "E804F996-ECB4-423F-895D-FD24E7B88CE4" : [
+      {
+        "date" : 753701153.832297,
+        "title" : "Untitled",
+        "transform" : [
+          0.9968556,
+          -1.982371e-09,
+          0.07923952,
+          0,
+          0.00092985097,
+          0.99993116,
+          -0.011697743,
+          0,
+          -0.07923409,
+          0.011734584,
+          0.99678695,
+          0,
+          -2.3009677,
+          -1.1383365,
+          15.200074,
+          1
+        ]
+      },
+      {
+        "date" : 753701150.148403,
+        "title" : "Untitled",
+        "transform" : [
+          0.9968556,
+          -1.982371e-09,
+          0.07923952,
+          0,
+          0.00092985097,
+          0.99993116,
+          -0.011697743,
+          0,
+          -0.07923409,
+          0.011734584,
+          0.99678695,
+          0,
+          -2.807861,
+          -1.3891072,
+          18.43909,
+          1
+        ]
+      },
+      {
+        "date" : 753701110.212527,
+        "title" : "Untitled",
+        "transform" : [
+          0.9968556,
+          -1.982371e-09,
+          0.07923952,
+          0,
+          0.00092985097,
+          0.99993116,
+          -0.011697743,
+          0,
+          -0.07923409,
+          0.011734584,
+          0.99678695,
+          0,
+          -16.716747,
+          0.19428948,
+          98.59264,
+          1
+        ]
+      },
+      {
+        "date" : 753700964.244446,
+        "title" : "Untitled",
+        "transform" : [
+          0.9971206,
+          -1.982371e-09,
+          0.07583184,
+          0,
+          0.00088986306,
+          0.99993116,
+          -0.011700853,
+          0,
+          -0.075826645,
+          0.011734584,
+          0.99705195,
+          0,
+          -5.631183,
+          0.06679501,
+          33.914726,
+          1
+        ]
+      },
+      {
+        "date" : 753699438.963171,
+        "title" : "Untitled",
+        "transform" : [
+          0.9976564,
+          -2.015628e-09,
+          -0.06842364,
+          0,
+          -0.013350854,
+          0.9807793,
+          -0.19466324,
+          0,
+          0.06710847,
+          0.1951205,
+          0.97848076,
+          0,
+          -0.124663964,
+          1.1024997,
+          5.862547,
+          1
+        ]
+      },
+      {
+        "date" : 753697979.436654,
+        "title" : "Untitled",
+        "transform" : [
+          0.9976564,
+          -2.015628e-09,
+          -0.06842364,
+          0,
+          -0.013350854,
+          0.9807793,
+          -0.19466324,
+          0,
+          0.06710847,
+          0.1951205,
+          0.97848076,
+          0,
+          0.07318884,
+          1.6777644,
+          8.747354,
+          1
+        ]
+      },
+      {
+        "date" : 753697947.3532,
+        "title" : "Untitled",
+        "transform" : [
+          0.9976564,
+          -2.015628e-09,
+          -0.06842364,
+          0,
+          -0.013350854,
+          0.9807793,
+          -0.19466324,
+          0,
+          0.06710847,
+          0.1951205,
+          0.97848076,
+          0,
+          0.47261778,
+          2.8391109,
+          14.571248,
+          1
+        ]
+      },
+      {
+        "date" : 753697926.357355,
+        "title" : "Untitled",
+        "transform" : [
+          0.99622047,
+          -2.015628e-09,
+          -0.0868612,
+          0,
+          -0.016948398,
+          0.9807792,
+          -0.19438303,
+          0,
+          0.08519164,
+          0.1951205,
+          0.9770724,
+          0,
+          -0.8714039,
+          1.5956244,
+          8.147981,
+          1
+        ]
+      },
+      {
+        "date" : 753697912.7817,
+        "title" : "Untitled",
+        "transform" : [
+          0.54651827,
+          -1.7559428e-09,
+          0.83744717,
+          0,
+          0.0034738402,
+          0.99999136,
+          -0.002267027,
+          0,
+          -0.83744,
+          0.0041481294,
+          0.54651356,
+          0,
+          -7.7268057,
+          0.029902667,
+          3.1607804,
+          1
+        ]
+      },
+      {
+        "date" : 753697801.584994,
+        "title" : "Untitled",
+        "transform" : [
+          0.5465183,
+          -1.8370401e-09,
+          0.8374472,
+          0,
+          0.0034738402,
+          0.9999915,
+          -0.0022670273,
+          0,
+          -0.83744,
+          0.0041480907,
+          0.54651356,
+          0,
+          -5.344624,
+          0.42580533,
+          8.349465,
+          1
+        ]
+      }
+    ]
+  },
+  "sceneViewportBackgroundColors" : {
+
+  }
+}
