+++ Cakey/composer/Sources/composer/composer.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+/// Bundle for the composer project
+public let composerBundle = Bundle.module
